acceptance: remove terraform and terrafarm
All the tests have been migrated to the new roachtest nightlies.

Release note: None

Closes #22646.
