--- pkg/acceptance/terraform/azure/.gitignore
@@ -1,1 +1,0 @@
-.terraform
